keibabook/requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.1.0
sqlalchemy>=2.0.0
schedule>=1.2.0
pytest>=7.0.0
pyyaml>=6.0
